renamed package to icounter.
