+++ icounter/const.py
