+++ icounter/fourier.py
